+++ src/test/java/de/kamillionlabs/hateoflux/linkbuilder/UriExpanderTest.java
@@ -1,9 +1,12 @@
 package de.kamillionlabs.hateoflux.linkbuilder;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
 
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -14,6 +17,50 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 class UriExpanderTest {
+    @ParameterizedTest
+    @CsvSource({ // This method under test removes the 3 query parameters page, size and sort.
+
+            // No query parameters
+            "'customer/34/orders', 'customer/34/orders'",
+
+            // Only other query parameter available
+            "'orders?userId=34', 'orders?userId=34'",
+
+            // No query parameters
+            "'customer/34/orders', 'customer/34/orders'",
+
+            // just sort present
+            "'customer/34/orders?sort=date,desc', 'customer/34/orders'",
+
+            // just size present
+            "'customer/34/orders?size=0', 'customer/34/orders'",
+
+            // just page present
+            "'customer/34/orders?page=0', 'customer/34/orders'",
+
+            // all the 3 present
+            "'customer/34/orders?page=0&size=20&sort=date,desc', 'customer/34/orders'",
+
+            // all the 3, sort twice and an extra query parameter at the front
+            "'orders?userId=34&page=0&size=20&sort=date,desc&sort=total,asc', 'orders?userId=34'",
+
+            // all the 3 and an extra query parameter at the front
+            "'orders?userId=34&page=0&size=20&sort=date,desc', 'orders?userId=34'",
+
+            // all the 3 and an extra query parameter at the back
+            "'orders?page=0&size=20&sort=date,desc&userId=34', 'orders?userId=34'",
+
+            // all the 3 and an extra query parameter in the middle
+            "'orders?page=0&size=20&userId=34&sort=date,desc', 'orders?userId=34'",
+
+            // all the 3 and an extra query parameter in the middle and with a full base URL
+            "'www.example.com/orders?page=0&size=20&userId=34&sort=date,desc', 'www.example.com/orders?userId=34'"
+    })
+    public void givenInputUrl_whenRemovePagingParameters_thenOutputUrl(String inputUrl, String expectedOutputUrl) {
+        String outputUrl = UriExpander.removePagingParameters(inputUrl);
+        assertThat(outputUrl).isEqualTo(expectedOutputUrl);
+    }
+
 
     @ParameterizedTest
     @CsvSource(delimiter = ';', value = {
@@ -191,13 +238,28 @@ class UriExpanderTest {
     }
 
     @Test
+    void givenExplodedQueryParameterWithSingleValue_whenExpandWithMap_thenCorrectUri() {
+        //GIVEN
+        String template = "/users{?keyWords*,limit}";
+        Map<String, Object> map = new LinkedHashMap<>();
+        map.put("keyWords", List.of("active"));
+        map.put("limit", 10);
+
+        //WHEN
+        String actual = UriExpander.expand(template, map);
+
+        //THEN
+        assertThat(actual).isEqualTo("/users?keyWords=active&limit=10");
+    }
+
+    @Test
     void givenNotExplodedQueryParameter_whenExpandWithMap_thenThrowException() {
         //GIVEN
         String template = "/users{?keyWords}";
         Map<String, Object> map = new LinkedHashMap<>();
         map.put("keyWords", List.of("active", "blue", "positive"));
         assertThatThrownBy(() -> UriExpander.expand(template, map)).isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("Detected a collection as value for a parameter, but parameter was not exploded in " +
+                .hasMessage("Detected a collection of values for a parameter, but parameter was not exploded in " +
                         "template (asterisk after parameter name e.g. {?var*}). " +
                         "Template was '/users{?keyWords}', parameters were {keyWords=[active, blue, positive]}");
     }
@@ -237,6 +299,20 @@ class UriExpanderTest {
         assertThatThrownBy(() -> UriExpander.expand(template, Map.of("key1", "value1", "key2", "value2")))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("Unknown parameters provided for URI template expansion.");
+    }
+
+
+    @Test
+    public void givenMultiValueMap_whenExpandWithMultiValueMap_thenPlaceholdersCorrectlySet() {
+        // GIVEN
+        MultiValueMap<String, String> queryParams = new LinkedMultiValueMap<>();
+        queryParams.add("someId", "54");
+
+        // WHEN
+        String expanded = UriExpander.expand("http://example.com{?someId}", queryParams);
+
+        //THEN
+        Assertions.assertThat(expanded).isEqualTo("http://example.com?someId=54");
     }