+++ src/test/java/de/kamillionlabs/hateoflux/model/hal/SerializationTest.java
@@ -273,4 +273,6 @@ public class SerializationTest {
                   }
                 """, actualJson, NON_EXTENSIBLE);
     }
+
+
 }