+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.0]
+
+### Changed
+
+* Allow expansion of URL even though values for paging are missing (are handled automatically by assemblers)
+
+### Fixed
+
+* Correctly process MultiValueMaps in expand() methods
+* Assemblers correctly build empty list wrappers if no list was provided
+
 ## [0.17.0]
 
 ### Added